feat: auction history fetcher interface 구현
